IoC регистрация объектов без параметров 01/06/2024
