+++ pom.xml
@@ -16,6 +16,13 @@
     <maven.compiler.target>21</maven.compiler.target>
   </properties>
 
+    <repositories>
+        <repository>
+            <id>spring-releases</id>
+            <url>https://repo.spring.io/release</url>
+        </repository>
+    </repositories>
+
   <dependencies>
 
       <dependency>
@@ -53,6 +60,8 @@
           <version>1.7.30</version>
           <scope>compile</scope>
       </dependency>
+
+
 
       <dependency>
           <groupId>ch.qos.logback</groupId>
