+++ src/main/java/iocModule2L1/CommandMove.java
@@ -1,4 +1,4 @@
-package iocModule2L9;
+package iocModule2L1;
 
 import exceptionModule1L3.ICommand;
