+++ src/main/java/iocModule2L1/DependencyResolver.java
@@ -1,0 +1,5 @@
+package iocModule2L1;
+
+public interface DependencyResolver<T> {
+    T resolve();
+}
