+++ src/main/java/iocModule2L1/MySupplier.java
@@ -1,0 +1,5 @@
+package iocModule2L1;
+
+public interface MySupplier<T, A> {
+    T get(A arg);
+}
